#115 First cut at making geonames an official jobs project
