+++ examples/geonames/build.gradle
@@ -4,11 +4,8 @@ plugins {
 
 repositories {
     jcenter()
-    mavenLocal()
 }
 
 dependencies {
-    compile "com.marklogic:marklogic-spring-batch:0.4.2"
-    testCompile "com.marklogic:marklogic-spring-batch-test:0.4.2"
-    runtime "com.marklogic:marklogic-spring-batch-jobs:0.4.2"
+    runtime "com.marklogic:marklogic-spring-batch-geonames:0.5.0"
 }
