--- examples/geonames/src/main/java/org/geonames/package-info.java
@@ -1,8 +1,0 @@
-@XmlSchema(
-	namespace="http://geonames.org",
-	elementFormDefault=XmlNsForm.QUALIFIED
-)
-package org.geonames;
-
-import javax.xml.bind.annotation.XmlSchema;
-import javax.xml.bind.annotation.XmlNsForm;
