--- examples/geonames/src/test/resources/logback.xml
@@ -1,23 +1,0 @@
-<configuration>
-
-  <statusListener class="ch.qos.logback.core.status.NopStatusListener" />
-
-  <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-    <encoder>
-      <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
-    </encoder>
-  </appender>
-
-  <root level="INFO">
-    <appender-ref ref="STDOUT" />
-  </root>
-
-  <logger name="com.marklogic" level="INFO" additivity="false">
-    <appender-ref ref="STDOUT" />
-  </logger>
-
-  <logger name="com.marklogic.client" level="WARN" additivity="false">
-    <appender-ref ref="STDOUT" />
-  </logger>
-
-</configuration>
