+++ jobs/geonames/src/main/resources/cities1.txt
@@ -1,0 +1,1 @@
+725993	Veliko Tŭrnovo	Veliko Turnovo	Tarnovo,Tarnowo,Ternovo,Tirnovo,Trnova,Trnovo,Turnovo,Tărnovo,Tărnowo,Tŭrnovo,Veliko Trnovo,Veliko Tarnovo,Veliko Tărnovo,Weliko Tarnowo,Wielkie Tyrnowo,vu~eriko・tarunovu~o,Велико Търново,ヴェリコ・タルノヴォ	43.08124	25.62904	P	PPLA	BG		62	VTR04			66217		207	Europe/Sofia	2007-07-01
