+++ settings.gradle
@@ -2,5 +2,6 @@ include "core"
 include "test"
 include "jobs"
 include "jobs:file-importers"
+include "jobs:geonames"
 include "jobs:shapefiles"
 include "jobs:sql"
